# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI20K Agent",
    "description": "AI Agent built with LangGraph",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/hotels": {
      "post": {
        "tags": [
          "admin-hotels"
        ],
        "summary": "Create Hotel",
        "operationId": "create_hotel_api_v1_admin_hotels_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHotelRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateHotelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "admin-hotels"
        ],
        "summary": "List Hotels",
        "operationId": "list_hotels_api_v1_admin_hotels_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "source",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "manual",
                "pipeline",
                "all"
              ],
              "type": "string",
              "default": "all",
              "title": "Source"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          },
          {
            "name": "embedding",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "embedded",
                "missing",
                "incomplete",
                "all"
              ],
              "type": "string",
              "default": "all",
              "title": "Embedding"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "json",
                "csv"
              ],
              "type": "string",
              "default": "json",
              "title": "Format"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HotelListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/embedding/summary": {
      "get": {
        "tags": [
          "admin-embedding"
        ],
        "summary": "Get Embedding Summary",
        "operationId": "get_embedding_summary_api_v1_admin_embedding_summary_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingSummaryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/embedding/missing": {
      "get": {
        "tags": [
          "admin-embedding"
        ],
        "summary": "Get Embedding Missing",
        "operationId": "get_embedding_missing_api_v1_admin_embedding_missing_get",
        "parameters": [
          {
            "name": "table",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "hotels",
                "rooms",
                "attractions"
              ],
              "type": "string",
              "title": "Table"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmbeddingMissingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/hotels/reembed": {
      "post": {
        "tags": [
          "admin-embedding"
        ],
        "summary": "Reembed Hotels",
        "operationId": "reembed_hotels_api_v1_admin_hotels_reembed_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReembedHotelsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReembedHotelsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateHotelRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "accommodation_type": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 50
              },
              {
                "type": "null"
              }
            ],
            "title": "Accommodation Type"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "star_rating": {
            "anyOf": [
              {
                "type": "number",
                "multipleOf": 0.5,
                "maximum": 5.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Star Rating"
          },
          "address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "destination_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Destination Id"
          },
          "city": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "check_in_time": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Check In Time"
          },
          "check_out_time": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Check Out Time"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreateHotelRequest",
        "description": "B2 (phase-08-hotel-create.md). `source_platform` is deliberately not a\nfield here -- accepting it from the client would let a caller claim an\nOTA origin for a hand-entered row, letting fake data into the ETL\nnamespace the pipeline trusts (see module docstring). It is always\n'manual', set server-side in `create_hotel`.\n\n`max_length`s below mirror the `hotels` column widths in\ndatabase_schema.sql (accommodation_type VARCHAR(50), address VARCHAR(500),\ncity VARCHAR(100)) so an over-long value 422s here instead of reaching\nPostgres as an unhandled `22001` (string too long)."
      },
      "CreateHotelResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source_platform": {
            "type": "string",
            "title": "Source Platform"
          },
          "source_hotel_id": {
            "type": "integer",
            "title": "Source Hotel Id"
          },
          "embedding_state": {
            "type": "string",
            "const": "missing",
            "title": "Embedding State"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_platform",
          "source_hotel_id",
          "embedding_state",
          "is_active"
        ],
        "title": "CreateHotelResponse"
      },
...
      "EmbeddingMissingItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "hotel_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hotel Name"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "EmbeddingMissingItem"
      },
      "EmbeddingMissingResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EmbeddingMissingItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "EmbeddingMissingResponse"
      },
      "EmbeddingSummaryResponse": {
        "properties": {
          "tables": {
            "items": {
              "$ref": "#/components/schemas/EmbeddingTableSummary"
            },
            "type": "array",
            "title": "Tables"
          },
          "total_missing": {
            "type": "integer",
            "title": "Total Missing"
          }
        },
        "type": "object",
        "required": [
          "tables",
          "total_missing"
        ],
        "title": "EmbeddingSummaryResponse"
      },
      "EmbeddingTableSummary": {
        "properties": {
          "table": {
            "type": "string",
            "enum": [
              "hotels",
              "rooms",
              "attractions"
            ],
            "title": "Table"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "embedded": {
            "type": "integer",
            "title": "Embedded"
          },
          "missing": {
            "type": "integer",
            "title": "Missing"
          }
        },
        "type": "object",
        "required": [
          "table",
          "label",
          "total",
          "embedded",
          "missing"
        ],
        "title": "EmbeddingTableSummary"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HotelListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/HotelRow"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "HotelListResponse"
      },
...
      "HotelRow": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "star_rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Star Rating"
          },
          "source_platform": {
            "type": "string",
            "title": "Source Platform"
          },
          "is_manual": {
            "type": "boolean",
            "title": "Is Manual"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "room_count": {
            "type": "integer",
            "title": "Room Count"
          },
          "hotel_embedded": {
            "type": "boolean",
            "title": "Hotel Embedded"
          },
          "rooms_missing_embedding": {
            "type": "integer",
            "title": "Rooms Missing Embedding"
          },
          "embedding_state": {
            "type": "string",
            "enum": [
              "embedded",
              "partial",
              "missing"
            ],
            "title": "Embedding State"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "source_platform",
          "is_manual",
          "is_active",
          "room_count",
          "hotel_embedded",
          "rooms_missing_embedding",
          "embedding_state"
        ],
        "title": "HotelRow"
      },
...
      "ReembedHotelsRequest": {
        "properties": {
          "hotel_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Hotel Ids"
          },
          "include_rooms": {
            "type": "boolean",
            "title": "Include Rooms",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "hotel_ids"
        ],
        "title": "ReembedHotelsRequest"
      },
      "ReembedHotelsResponse": {
        "properties": {
          "cleared_hotels": {
            "type": "integer",
            "title": "Cleared Hotels"
          },
          "cleared_rooms": {
            "type": "integer",
            "title": "Cleared Rooms"
          },
          "dag_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dag Run Id"
          },
          "queued": {
            "type": "boolean",
            "title": "Queued"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "cleared_hotels",
          "cleared_rooms",
          "queued"
        ],
        "title": "ReembedHotelsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}